+++ tx-models/src/main/java/qualitynotification/base/response/QualityNotificationMessageResponse.java
@@ -22,13 +22,17 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.RequiredArgsConstructor;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
 @Data
 @Builder
+@AllArgsConstructor
+@RequiredArgsConstructor
 public class QualityNotificationMessageResponse {
 
     private String id;